Fix limitation of 3D path length #33.
The main reason the Path and Polygon have limited number of points to be drawn is
that vertexBuffer indices are UNSIGNED_SHORT type and thus
unique index values are limited due to "Max Short" value.

Add IntBufferObject, make Path use it as drawState.elementBuffer.
Change interiorElements, outlineElements and verticalElements type to
Int.